+++ profiles/serializers/category_serializer.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from ..models import Category
+
+
+class CategorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Category
+        fields = "__all__"
