+++ profiles/serializers/profile_detail_serializers.py
@@ -76,6 +76,72 @@ class ProfileDetailSerializer(serializers.ModelSerializer):
         return get_regions_ukr_names_as_string(obj)
 
 
+class ProfileDetailSerializer(serializers.ModelSerializer):
+    activities = ActivitySerializer(many=True, read_only=True)
+    categories = CategorySerializer(many=True, read_only=True)
+    is_saved = serializers.SerializerMethodField()
+    banner_image = serializers.ImageField(required=False)
+    regions = RegionSerializer(many=True, read_only=True)
+    regions_ukr_display = serializers.SerializerMethodField()
+
+    class Meta:
+        model = Profile
+        fields = (
+            "id",
+            "official_name",
+            "regions",
+            "regions_ukr_display",
+            "common_info",
+            "edrpou",
+            "rnokpp",
+            "founded",
+            "address",
+            "startup_idea",
+            "name",
+            "is_registered",
+            "is_startup",
+            "is_fop",
+            "categories",
+            "activities",
+            "service_info",
+            "product_info",
+            "banner_image",
+            "logo_image",
+            "is_saved",
+        )
+        read_only_fields = (
+            "id",
+            "official_name",
+            "regions",
+            "regions_ukr_display",
+            "common_info",
+            "edrpou",
+            "rnokpp",
+            "founded",
+            "address",
+            "startup_idea",
+            "name",
+            "is_registered",
+            "is_startup",
+            "is_fop",
+            "categories",
+            "activities",
+            "service_info",
+            "product_info",
+            "banner_image",
+            "logo_image",
+        )
+
+    def get_is_saved(self, obj) -> bool:
+        user = self.context["request"].user
+        if user.is_authenticated:
+            return obj.pk in self.context["saved_companies_pk"]
+        return False
+
+    def get_regions_ukr_display(self, obj) -> str:
+        return get_regions_ukr_names_as_string(obj)
+
+
 class ProfileOwnerDetailViewSerializer(serializers.ModelSerializer):
     activities = ActivitySerializer(many=True, read_only=True)
     categories = CategorySerializer(many=True, read_only=True)
