+++ profiles/urls.py
@@ -4,17 +4,15 @@ from .views.saved_companys_views import (
     SavedCompaniesDestroy,
 )
 from .views.views import (
-    CategoryList,
     ActivityList,
     RegionList,
-    CategoryDetail,
     ActivityDetail,
     RegionDetail,
-    ProfileViewCreate,
 )
-
+from .views.viewed_company_views import ProfileViewCreate
 from .views.profile_list_views import ProfileList
 from .views.profile_detail_views import ProfileDetail
+from .views.category_views import CategoryList, CategoryDetail
 
 
 app_name = "profiles"
