+++ profiles/views/category_views.py
@@ -1,0 +1,20 @@
+from rest_framework.generics import (
+    ListCreateAPIView,
+    RetrieveUpdateDestroyAPIView,
+)
+from rest_framework.permissions import IsAdminUser
+from ..permissions import RequestIsReadOnly
+from ..models import Category
+from ..serializers.category_serializer import CategorySerializer
+
+
+class CategoryList(ListCreateAPIView):
+    serializer_class = CategorySerializer
+    permission_classes = (RequestIsReadOnly | IsAdminUser,)
+    queryset = Category.objects.all()
+
+
+class CategoryDetail(RetrieveUpdateDestroyAPIView):
+    serializer_class = CategorySerializer
+    permission_classes = (IsAdminUser,)
+    queryset = Category.objects.all()
